refactor: ProjectMedia
